% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bernoulliWordStatistics.R
\name{bernoulliWordStatistics}
\alias{bernoulliWordStatistics}
\title{Bernoulli word statistcs}
\usage{
bernoulliWordStatistics(seq, word_length)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to calculate the word statistics (mean and variance) using the Bernoulli model.
}
\examples{
cat_function()
}
\keyword{cats}

