% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateMotifs.R
\name{calculateMotifs}
\alias{calculateMotifs}
\title{Motif finding}
\usage{
calculateMotifs(kmerWeightVector, searchSeq = NULL, max.test = 5,
  motifNumber = 3, extend = 2, format = "consensus",
  unknownWeight = 0.01, plotCluster = FALSE, max.mismatch = 1, ...)
}
\arguments{
\item{seq}{The input DNA sequence}

\item{word_length}{The length of the words}
}
\value{
returns the following in the form of a list \item{var}{the variance of the words}
\item{mu}{the mean of the words}\item{N_w}{the word count}
}
\description{
Calculate the motifs using word statistics.
}
\examples{
library(Biostrings)
seq <- readDNAStringSet(filePath,format="fasta")
word_length <- 2
t <- bernoulliWordStatistics(seq,word_length)
}
\keyword{Bernoulli,}
\keyword{finding}
\keyword{motif}
\keyword{statistics,}
\keyword{word}

