% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateOverlapIndicator.R
\name{calculateOverlapIndicator}
\alias{calculateOverlapIndicator}
\title{Calculate overlap between two words}
\usage{
calculateOverlapIndicator(word1, word2)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to calculate the overlap between two words
}
\examples{
cat_function()
}
\keyword{cats}

