\name{wordStats-package}
\alias{wordStats-package}
\alias{wordStats}
\docType{package}
\title{
Calculate DNA motifs using word statistics 
}
\description{
Given a DNA sequence, this package calculates motifs using word statistics. The word statistics (mean and variance) are computed using Bernoulli or Markov Model
}
\details{
\tabular{ll}{
Package: \tab wordStats\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-04\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
