% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateOverlapIndicator.R
\name{calculateOverlapIndicator}
\alias{calculateOverlapIndicator}
\title{Calculate word overlap}
\usage{
calculateOverlapIndicator(word1, word2)
}
\arguments{
\item{word1}{The first word}

\item{word2}{The second word}
}
\value{
\item {Outputs the overlap between the two words. The size of the overlap is the same as the input word with the longest length}
}
\description{
Given two words, calculate overlap between them by shifting the second word over the first one, one letter at a time over
}
\examples{
word1 = 'ACGA'
word2 = 'TCGC'
overlap <-calculateOverlapIndicator(word1,word2)
}
\keyword{overlap}
\keyword{word,}

